requirements.txt
# Django REST API Requirements for OneStopRadio
# Core Django & REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database & Authentication
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File uploads & Storage
Pillow==10.1.0
django-storages==1.14.2

# Social Media Integration  
requests==2.31.0
python-social-auth[django]==0.12.0

# API Documentation
drf-spectacular==0.26.5

# Development tools
django-debug-toolbar==4.2.0
python-decouple==3.8

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3